captioning-app/backend/requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
Pillow
python-multipart
opencv-python
matplotlib
sqlalchemy
